Move tools' HDF5 module up one level
And adjust tool dependencies.
