+++ pyproject.toml
@@ -64,9 +64,8 @@ clients = [
     "textual",
 ]
 tools = [
+    "caterva2[hdf5]",
     "blosc2",
-    "h5py",
-    "hdf5plugin",
     "msgpack",
 ]
 tests = [
